Edited from 3 to 4.2.1:
Made proofreading changes to the English version,
for the function prologue/epilogue chapter,
until the end of 4.2.1.

One small change to content is saying, in the intro to 4,
"Nevertheless ARM not only has instructions supporting descending stacks
but also ascending stacks.", since the descending stack is standard,
but the latter is not. I believe this also makes it consistent with the
Russian version.